Backend/requirements.txt
fastapi
uvicorn
torch
ultralytics
pillow
python-multipart
